refactor: 分离类
